Got Rid of Unnecessary Room File Check for isRoom() Function
Deleted a portion of the code in the function isRoom(), which returns
true if the BoardCell is a room. It was previously checking to make sure
the initial was in the file ClueRooms.txt and then returning if the
initial was not "W". The file check is useless because we are not always
using the file "ClueRooms.txt", and the initial the BoardCell gets comes
from whatever file we are using, so it is a redundant check.