Got Rid of Unnecessary Variable in loadRoomConfig Main Loop, currentLine
The currentLine variable was being used as an intermediate variable that
holds the line that is read from the file, before it is being split up
into its components using the split() function. Instead of having this
variable, I changed it to just .split the next line as soon as it is
read in.